+++ assert/assertion.go
@@ -19,11 +19,10 @@ type OptionalAssertion interface {
 	// End is called once the test is complete.
 	End()
 
-	// Ok returns true if the assertion passed.
+	// TryOk returns true if the assertion passed.
 	//
-	// If asserted is false, the assertion was a no-op and the value of pass is
-	// meaningless.
-	Ok() (ok bool, asserted bool)
+	// If asserted is false, the assertion was a no-op and ok is meaningless.
+	TryOk() (ok bool, asserted bool)
 
 	// BuildReport generates a report about the assertion.
 	//
@@ -38,8 +37,8 @@ type OptionalAssertion interface {
 type Assertion interface {
 	OptionalAssertion
 
-	// MustOk returns true if the assertion passed.
-	MustOk() bool
+	// Ok returns true if the assertion passed.
+	Ok() bool
 }
 
 // Nothing is an "optional assertion" that always passes and does not build an
@@ -51,7 +50,7 @@ type noopAssertion struct{}
 func (noopAssertion) Notify(fact.Fact)         {}
 func (noopAssertion) Begin(compare.Comparator) {}
 func (noopAssertion) End()                     {}
-func (noopAssertion) Ok() (bool, bool)         { return false, false }
+func (noopAssertion) TryOk() (bool, bool)      { return false, false }
 func (noopAssertion) BuildReport(bool, bool, render.Renderer) *Report {
 	panic("not implemented")
 }