+++ test_test.go
@@ -167,8 +167,8 @@ type noopAssertion struct{}
 
 func (noopAssertion) Begin(compare.Comparator) {}
 func (noopAssertion) End()                     {}
-func (noopAssertion) Ok() (bool, bool)         { return true, true }
-func (noopAssertion) MustOk() bool             { return true }
+func (noopAssertion) TryOk() (bool, bool)      { return true, true }
+func (noopAssertion) Ok() bool                 { return true }
 func (noopAssertion) Notify(fact.Fact)         {}
 
 func (noopAssertion) BuildReport(ok, verbose bool, r render.Renderer) *assert.Report {